feat: support reset action info
Add a ResetActionInfo route and link it from the ComputerSysetm Reset
action.

This requires manually modifying generated code from the Redfish OpenAPI
definition. That definition seemingly does not account for generic
fields like ActionInfo, so while it generates the API types for them, it
doesn't use them in any objects.